+++ src/sql/src/catalog.rs
@@ -13,10 +13,11 @@
 
 use std::fmt;
 use std::path::PathBuf;
-use std::time::Instant;
+use std::time::{Duration, Instant};
 use std::{error::Error, unimplemented};
 
 use chrono::{DateTime, Utc, MIN_DATETIME};
+use dataflow_types::SourceConnector;
 use lazy_static::lazy_static;
 
 use build_info::{BuildInfo, DUMMY_BUILD_INFO};
@@ -163,6 +164,8 @@ pub struct CatalogConfig {
     pub build_info: &'static BuildInfo,
     /// The number of worker in use by the server.
     pub num_workers: usize,
+    /// Default timestamp frequency for CREATE SOURCE
+    pub timestamp_frequency: Duration,
 }
 
 /// A database in a [`Catalog`].
@@ -217,6 +220,12 @@ pub trait CatalogItem {
     /// If the catalog item is not of a type that produces functions (i.e.,
     /// anything other than a function), it returns an error.
     fn func(&self) -> Result<&'static Func, CatalogError>;
+
+    /// Returns the resolved source connector.
+    ///
+    /// If the catalog item is not of a type that contains a `SourceConnector`
+    /// (i.e., anything other than sources), it returns an error.
+    fn source_connector(&self) -> Result<&SourceConnector, CatalogError>;
 
     /// Returns the type of the catalog item.
     fn item_type(&self) -> CatalogItemType;
@@ -290,6 +299,8 @@ pub enum CatalogError {
     UnknownItem(String),
     /// Unknown function.
     UnknownFunction(String),
+    /// Unknown source.
+    UnknownSource(String),
     /// Invalid attempt to depend on a non-dependable item.
     InvalidDependency {
         /// The invalid item's name.
@@ -304,6 +315,7 @@ impl fmt::Display for CatalogError {
         match self {
             Self::UnknownDatabase(name) => write!(f, "unknown database '{}'", name),
             Self::UnknownFunction(name) => write!(f, "function \"{}\" does not exist", name),
+            Self::UnknownSource(name) => write!(f, "source \"{}\" does not exist", name),
             Self::UnknownSchema(name) => write!(f, "unknown schema '{}'", name),
             Self::UnknownRole(name) => write!(f, "unknown role '{}'", name),
             Self::UnknownItem(name) => write!(f, "unknown catalog item '{}'", name),
@@ -342,6 +354,7 @@ lazy_static! {
         cache_directory: None,
         build_info: &DUMMY_BUILD_INFO,
         num_workers: 0,
+        timestamp_frequency: Duration::from_secs(1)
     };
 }