+++ doc/user/content/release-notes.md
@@ -48,6 +48,9 @@ Wrap your release notes at the 80 character mark.
 
 {{% version-header v0.7.2 %}}
 
+- Record Kafka Consumer metrics in the `mz_kafka_consumer_statistics` system
+  table. Enabled by default for all Kafka sources.
+
 - Add the [`jsonb_object_agg`](/sql/functions/jsonb_object_agg) function to
   aggregate rows into a JSON object.